+++ src/app/Data/services/charts.ts
@@ -1,6 +1,0 @@
-import { Chart } from 'chart.js/auto';
-
-class Charts {
-  private constructor(chart: Chart) {
-  }
-}